main/postmarketos-initramfs: add support for immutable booting
This adds support for booting immutable pmOS images. The new logic here
is triggered on the presence of "usrhash=" on the kernel cmdline, this
param is used by dm-verity and makes no sense to have there unless there
are RO verified partitions (e.g. an immutable /usr or something). Once
this is triggered, a few new things can happen:

"First boot" - disk with booted ESP is detected, detects the correct
usr partition to verify/mount using usrhash, and runs f0rmz to collect
username, password, (optional) fde passphrase and (optional) hostname.
Then systemd-repart is used to create the rest of the partition layout
(rootfs, slot B usr partition stuff), and (optionally) encrypts the
rootfs. Next it runs systemd-sysusers and systemd-firstboot on the new
root partition to create the user account and set hostname and some
other tings, mounts the booted usr partition into it and then boot
continues (e.g. to switch_root)

"Normal boot" - disk with booted ESP is detected, this disk is searched
for the appropriate root partition type for the booted arch and FDE
unlocked if necessary. usrhash is used to verify and mount the usr
partition. Then boot continues (e.g. to switch_root)

"Factory reset" - A reset condition is detected by looking at a specific
EFI variable set on the previous boot, and systemd-repart is run to
erase the rootfs. Then the initramfs continues with the "First boot"
flow above, where the user can continue to configure a new install
OR shut down the device to leave it in an unconfigured state to be
continued later.

It's important to note that finding the root partition is implemented
in a different way than how the initramfs finds the root partition for
mutable booting. I attempted to lean on the Discoverable Partition Spec,
which depends on an EFI variable and partition type UUIDs to locate
appropriate disks and partitions to search.

[ci:ignore-count]

Signed-off-by: Clayton Craft <craftyguy@postmarketos.org>

Signed-off-by: Clayton Craft <craftyguy@postmarketos.org>