ci: fix test_providers.py to not fail if provider_priority is 0
As per the apk-tools and abuild source code, 0 is a perfectly valid integer
for provider_priority, but our Python tests currently fail with a priority of 0
because 0 and None are the same in the if statement that checks priority.

Changes the if statement that checks priority to fail if the priority is
explicitly None.

Signed-off-by: Aster Boese <asterboese@mailbox.org>