kconfigcheck: Add recommended kernel hardening options
This adds the recommended hardening options from
kernel/configs/hardening.config to the hardening category in
kconfigcheck.

I have only added an explicit version requirement if the config was
changed/added in the last 2 years, since otherwise it would become
overly verbose. The category is not meant to be enabled by downstream
kernels anyways.

Signed-off-by: Jens Reidel <adrian@postmarketos.org>